Add Duration and Instant classes
These allow time representation and arithmetics with some protection from
over/underflow. I decided to skip std::chrono so that the definition of a
time point (Instant) is not tied to a particular clock but simply defined
to be relative to the Unix epoch. Also, microsecond precision is good enough
for everyone, we don't need the full flexibility that 'std::chrono` provides,
and can avoid some of its complexity. This should then also make mocking
time in unit tests a bit easier.