Add basic JSON loader functionality
This will be used to load a scenario and game state from JSON files, and
also to transfer them between servers and clients. The complex part of it
is the include functionality which is used to be able to e.g. store
definitions in separate files that can be referenced from the main scenario
file (including custom scenario-specific overrides).

The information about what was included from where is also preserved so that
when writing the information back to file, not the whole tree is dumped but
the same include references are replicated as from when the file was loaded.
The save-back functionality is not implemented yet, this will follow in a
subsequent PR, with hopefully only small changes to the loader still needed
for it.