Add Collection and some utilities around it
A "collection" is basically a std::map<std::string, T>. The idea is to keep
the object IDs part of the game state instead of just during the
(de)serialization. This will simplify things considerably when operations
on the game state need to be (de)serialized, e.g. when manipulations of the
game state are being communicated over the network.

The object ID itself is not contained in the objects, but only stored in
the key of the map (collection). References to an object in a collection are
stored as an iterator into the map so that the object ID is available if
needed.