Allow non-const dereferincing of def::Ref (#24)
Allow non-const dereferincing of def::Ref

Instead, make it non-copyable, so that one cannot get a non-const Ref from
a const Ref. This makes the Refs a bit easier to use, but one needs to be
a bit more careful when passing them around.

Then, have CollectionLoader and CollectionSaver pass def::Ref instead of a normal reference
to the child handler. Then the child handler has access to the object ID and the ref in case
it needs to be read or set somewhere.